# Conflicts:
#   database/populate.sql

+++ app/Http/Controllers/AuctionController.php
@@ -97,7 +97,28 @@ class AuctionController extends Controller
     public function delete($id)
     {
       $auction = Auction::find($id);
+      $this->authorize('delete', $auction);
       $auction->delete();
       return redirect('/');
     }
+
+    public function listBids($id)
+    {
+      $auction = Auction::find($id);
+      $bids = $auction->bids()->orderBy('amount', 'desc')->get();
+      return view('pages.bids', ['bids' => $bids]);
+    }
+
+    
+    // search using tsvectors
+    public function ftsSearch(Request $request)
+    { 
+      $search = $request->get('search');
+      $formattedSearch = str_replace(' ', '|', $search);
+      $auctions = Auction::whereRaw("tsvectors @@ to_tsquery('english', ?)", [$formattedSearch])->get();
+    
+      //$auctions = Auction::where('name', 'LIKE', '%' . $search . '%')->get();
+      return view('pages.auctionsListing', ['auctions' => $auctions]);
+    }
+
 }
